--- src/pages/Homepage/components/imgExports.jsx
@@ -1,6 +1,0 @@
-import boardImg from "../../../assets/board-brown-white(reduced).webp";
-import pieceImg from "../../../assets/chess-black-brown-pieces(reduced).webp";
-import bookImg from "../../../assets/chess-fianchetto-book(reduced).webp";
-import clockImg from "../../../assets/chess-clock(reduced).webp";
-
-export { boardImg, pieceImg, bookImg, clockImg };
